move big int to byte slice conversion functions to separate file
