+++ runtime/interpreter/big.go
@@ -1,0 +1,57 @@
+package interpreter
+
+import (
+	"math/big"
+
+	"github.com/onflow/cadence/runtime/errors"
+)
+
+func SignedBigIntToBigEndianBytes(bigInt *big.Int) []byte {
+
+	switch bigInt.Sign() {
+	case -1:
+		// Encode as two's complement
+		twosComplement := new(big.Int).Neg(bigInt)
+		twosComplement.Sub(twosComplement, big.NewInt(1))
+		bytes := twosComplement.Bytes()
+		for i := range bytes {
+			bytes[i] ^= 0xff
+		}
+		// Pad with 0xFF to prevent misinterpretation as positive
+		if len(bytes) == 0 || bytes[0]&0x80 == 0 {
+			return append([]byte{0xff}, bytes...)
+		}
+		return bytes
+
+	case 0:
+		return []byte{0}
+
+	case 1:
+		bytes := bigInt.Bytes()
+		// Pad with 0x0 to prevent misinterpretation as negative
+		if len(bytes) > 0 && bytes[0]&0x80 != 0 {
+			return append([]byte{0x0}, bytes...)
+		}
+		return bytes
+
+	default:
+		panic(errors.NewUnreachableError())
+	}
+}
+
+func UnsignedBigIntToBigEndianBytes(bigInt *big.Int) []byte {
+
+	switch bigInt.Sign() {
+	case -1:
+		panic(errors.NewUnreachableError())
+
+	case 0:
+		return []byte{0}
+
+	case 1:
+		return bigInt.Bytes()
+
+	default:
+		panic(errors.NewUnreachableError())
+	}
+}
